RUM-2925 Add `BacktraceReport` API to capture thread backtrace
