+++ DatadogInternal/Sources/FeatureAPIs/CrashReporting/BinaryImage.swift
@@ -6,39 +6,6 @@
 
 import Foundation
 
-/// Unsymbolicated stack trace thread.
-public struct DDThread: Codable {
-    /// The name of the thread, e.g. `"Thread 0"`
-    public let name: String
-    /// Unsymbolicated stack trace of the crash.
-    public let stack: String
-    /// If the thread was halted.
-    public let crashed: Bool
-    /// Thread state (CPU registers dump), only available for halted thread.
-    public let state: String?
-
-    public init(
-        name: String,
-        stack: String,
-        crashed: Bool,
-        state: String?
-    ) {
-        self.name = name
-        self.stack = stack
-        self.crashed = crashed
-        self.state = state
-    }
-
-    // MARK: - Encoding
-
-    enum CodingKeys: String, CodingKey {
-        case name = "name"
-        case stack = "stack"
-        case crashed = "crashed"
-        case state = "state"
-    }
-}
-
 /// Binary Image referenced in frames from `DDThread`.
 public struct BinaryImage: Codable {
     public let libraryName: String
